expose cs printing
