+++ src/lib/provable-context.ts
@@ -2,6 +2,7 @@ import { Context } from './global-context.js';
 import { Gate, JsonGate, Snarky } from '../snarky.js';
 import { parseHexString } from '../bindings/crypto/bigint-helpers.js';
 import { prettifyStacktrace } from './errors.js';
+import { Fp } from '../bindings/crypto/finite_field.js';
 
 // internal API
 export {
@@ -17,6 +18,7 @@ export {
   inCompile,
   inCompileMode,
   gatesFromJson,
+  printGates,
 };
 
 // global circuit-related context
@@ -94,7 +96,16 @@ function constraintSystem<T>(f: () => T) {
       result = f();
     });
     let { gates, publicInputSize } = gatesFromJson(json);
-    return { rows, digest, result: result! as T, gates, publicInputSize };
+    return {
+      rows,
+      digest,
+      result: result! as T,
+      gates,
+      publicInputSize,
+      print() {
+        printGates(gates);
+      },
+    };
   } catch (error) {
     throw prettifyStacktrace(error);
   } finally {
@@ -106,8 +117,62 @@ function constraintSystem<T>(f: () => T) {
 
 function gatesFromJson(cs: { gates: JsonGate[]; public_input_size: number }) {
   let gates: Gate[] = cs.gates.map(({ typ, wires, coeffs: hexCoeffs }) => {
-    let coeffs = hexCoeffs.map(hex => parseHexString(hex).toString());
+    let coeffs = hexCoeffs.map((hex) => parseHexString(hex).toString());
     return { type: typ, wires, coeffs };
   });
   return { publicInputSize: cs.public_input_size, gates };
+}
+
+// print a constraint system
+
+function printGates(gates: Gate[]) {
+  for (let i = 0, n = gates.length; i < n; i++) {
+    let { type, wires, coeffs } = gates[i];
+    console.log(
+      i.toString().padEnd(4, ' '),
+      type.padEnd(15, ' '),
+      coeffsToPretty(type, coeffs).padEnd(30, ' '),
+      wiresToPretty(wires, i)
+    );
+  }
+  console.log();
+}
+
+let minusRange = Fp.modulus - (1n << 64n);
+
+function coeffsToPretty(type: Gate['type'], coeffs: Gate['coeffs']): string {
+  if (coeffs.length === 0) return '';
+  if (type === 'Generic' && coeffs.length > 5) {
+    let first = coeffsToPretty(type, coeffs.slice(0, 5));
+    let second = coeffsToPretty(type, coeffs.slice(5));
+    return `${first} ${second}`;
+  }
+  if (type === 'Poseidon' && coeffs.length > 3) {
+    return `${coeffsToPretty(type, coeffs.slice(0, 3)).slice(0, -1)} ...]`;
+  }
+  let str = coeffs
+    .map((c) => {
+      let c0 = BigInt(c);
+      if (c0 > minusRange) c0 -= Fp.modulus;
+      let cStr = c0.toString();
+      if (cStr.length > 4) return `${cStr.slice(0, 4)}..`;
+      return cStr;
+    })
+    .join(' ');
+  return `[${str}]`;
+}
+
+function wiresToPretty(wires: Gate['wires'], row: number) {
+  let strWires: string[] = [];
+  let n = wires.length;
+  for (let col = 0; col < n; col++) {
+    let wire = wires[col];
+    if (wire.row === row && wire.col === col) continue;
+    if (wire.row === row) {
+      strWires.push(`${col}->${wire.col}`);
+    } else {
+      strWires.push(`${col}->(${wire.row},${wire.col})`);
+    }
+  }
+  return strWires.join(', ');
 }
